requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.19.0
scipy>=1.7.0
PyYAML>=6.0
